# Improve performance and reduce allocations in Plan.