--- a/src/Ninject/Planning/Plan.cs
+++ b/src/Ninject/Planning/Plan.cs
@@ -96,7 +96,13 @@ namespace Ninject.Planning
         public IEnumerable<TDirective> GetAll<TDirective>()
             where TDirective : IDirective
         {
-            return this.Directives.OfType<TDirective>();
+            foreach (var directive in this.Directives)
+            {
+                if (directive is TDirective tdir)
+                {
+                    yield return tdir;
+                }
+            }
         }
     }
 }