* Improves performance and reduces allocations in Plan.
* Add unit tests for Plan.

**Before:**

|                    Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|  HasDirective_Match_First |  77.12 ns | 0.9070 ns | 0.8484 ns |      0.0228 |           - |           - |                96 B |
| HasDirective_Match_Middle | 169.68 ns | 1.1336 ns | 1.0603 ns |      0.0226 |           - |           - |                96 B |
|   HasDirective_Match_Last | 240.03 ns | 0.8890 ns | 0.7881 ns |      0.0224 |           - |           - |                96 B |
|      HasDirective_NoMatch | 256.20 ns | 1.8254 ns | 1.7075 ns |      0.0224 |           - |           - |                96 B |
|        GetOne_Match_First | 277.86 ns | 1.1517 ns | 1.0773 ns |      0.0224 |           - |           - |                96 B |
|       GetOne_Match_Middle | 281.71 ns | 1.6616 ns | 1.4730 ns |      0.0224 |           - |           - |                96 B |
|         GetOne_Match_Last | 266.67 ns | 2.2552 ns | 2.1095 ns |      0.0224 |           - |           - |                96 B |
|            GetOne_NoMatch | 260.81 ns | 1.4024 ns | 1.3118 ns |      0.0224 |           - |           - |                96 B |
|              GetAll_Match |  25.47 ns | 0.2412 ns | 0.2256 ns |      0.0133 |           - |           - |                56 B |
|            GetAll_NoMatch |  25.98 ns | 0.3688 ns | 0.3450 ns |      0.0133 |           - |           - |                56 B |


**After:**

|                    Method |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|  HasDirective_Match_First |  39.620 ns | 0.2334 ns | 0.2069 ns |      0.0209 |           - |           - |                88 B |
| HasDirective_Match_Middle | 121.500 ns | 1.5366 ns | 1.3622 ns |      0.0207 |           - |           - |                88 B |
|   HasDirective_Match_Last | 189.981 ns | 0.7099 ns | 0.6640 ns |      0.0207 |           - |           - |                88 B |
|      HasDirective_NoMatch | 209.746 ns | 2.0663 ns | 1.9329 ns |      0.0207 |           - |           - |                88 B |
|        GetOne_Match_First | 226.161 ns | 1.0210 ns | 0.8526 ns |      0.0207 |           - |           - |                88 B |
|       GetOne_Match_Middle | 226.589 ns | 0.6180 ns | 0.5781 ns |      0.0207 |           - |           - |                88 B |
|         GetOne_Match_Last | 218.492 ns | 2.5944 ns | 2.4268 ns |      0.0207 |           - |           - |                88 B |
|            GetOne_NoMatch | 219.027 ns | 0.7464 ns | 0.6982 ns |      0.0207 |           - |           - |                88 B |
|              GetAll_Match |   5.787 ns | 0.0383 ns | 0.0358 ns |      0.0114 |           - |           - |                48 B |
|            GetAll_NoMatch |   5.788 ns | 0.0583 ns | 0.0545 ns |      0.0114 |           - |           - |                48 B |